feat: 最新表と過去表の構成から表示形式(A表・B表)に対応できるように構成を変更
